Add files via upload
Explanation :

01. Package Declaration:

Copy code :
                     package BasicPrograms;

This line declares that the class belongs to the BasicPrograms package.

02. Class Declaration:

Copy code :
                     public class LcmOfTwoNumbers 

This line declares a public class named LcmOfTwoNumbers.

03. Main Method:

Copy code :
                     public static void main(String args[])

This is the main method where the program execution begins. The String args[] parameter allows the program to accept command-line arguments.

04. Variable Initialization:

Copy code :
                     int R = 45;
                     int S = 78;
                     int gcd = 1;

Here, two integers R and S are declared and initialized with the values 45 and 78, respectively. Another integer gcd is initialized to 1, which will be used to store the Greatest Common Divisor.

05. Finding the GCD:

Copy code :
                     for(int i = 1; i <= R && i <= S; ++i)
                     {
                         if(R % i == 0 && S % i == 0)
                        gcd = i;
                     }

This for loop runs from 1 to the minimum of R and S. For each value of i:

              * It checks if i is a divisor of both R and S using the modulus operator (%).
              * If true, it updates gcd to i. By the end of the loop, gcd will hold the largest number that divides both R and S.

06. Calculating the LCM:

Copy code : 
                     int lcm = (R * S) / gcd;

The Least Common Multiple (LCM) is calculated using the formula (R * S) / gcd.

07. Printing the Result:

Copy code :
                      System.out.println("The LCM of R and S is : " + lcm);

This line prints the LCM of the two numbers R and S.

Summary:

The program calculates the LCM of two given numbers by first finding their GCD using a loop and then applying the LCM formula (R * S) / gcd. The result is then printed to the console.