Add files via upload
Explanation :

This Java program finds the largest among three numbers (F, G, and H) :

01. Package and Class Declaration:

 code :
             package BasicPrograms;

             public class LargestOfThreeNumbers 
             {

The program is declared inside a package named BasicPrograms. It contains a class named LargestOfThreeNumbers.

02. Non-Static Method Declaration:

code :
           public void methodName(int F,double G,int H)    
           {
 	         // Method Body
           }

This is a non-static method named methodName. It takes three parameters: F (an integer), G (a double), and H (an integer). It doesn't return anything (void). This method is responsible for determining the largest among the three numbers and printing the result.

03. Logic for Finding the Largest Number:

code :
            if (F>=G && F>=H)
 	    System.out.println(F+ " Is The Largest Number.");
            else if (G>=F && G>=H)
     	    System.out.println(G+ " Is The Largest number.");
            else
            System.out.println(H+ " Is The Largest number.");
	 
The program checks conditions to find the largest number among F, G, and H.

          * If F is greater than or equal to both G and H, then F is the largest.
          * If G is greater than or equal to both F and H, then G is the largest.
          * If the above conditions are not met, then H is the largest.

04. Main Method:

 code :
             public static void main(String args[]) 
             {
 	          LargestOfThreeNumbers ggg = new LargestOfThreeNumbers();
 	          ggg.methodName(32,82.5,44);  
             }

The main method is the entry point of the program.
Inside the main method, an instance of the LargestOfThreeNumbers class is created.

The methodName method of this instance is then called with three arguments: 32, 82.5, and 44.

05. Output:
              The program prints the largest number among the three provided numbers.

              This program effectively determines the largest number among three given numbers using conditional statements.